:bulb: Feat: 앨범 상세 데이터 받아서 뿌리기(#56)
