--- src/pages/Album/data.ts
@@ -1,39 +1,0 @@
-import FeedImg1 from '../../asset/image/feed-test/1.jpg';
-import FeedImg2 from '../../asset/image/feed-test/2.jpg';
-import FeedImg3 from '../../asset/image/feed-test/3.jpg';
-import FeedImg4 from '../../asset/image/feed-test/4.png';
-import FeedImg5 from '../../asset/image/feed-test/5.jpg';
-import FeedImg6 from '../../asset/image/feed-test/6.jpg';
-import FeedImg7 from '../../asset/image/feed-test/7.jpg';
-import FeedImg8 from '../../asset/image/feed-test/8.jpg';
-import FeedImg9 from '../../asset/image/feed-test/9.jpg';
-import FeedImg10 from '../../asset/image/feed-test/10.jpg';
-import FeedImg11 from '../../asset/image/feed-test/11.jpg';
-import FeedImg12 from '../../asset/image/feed-test/12.jpg';
-import FeedImg13 from '../../asset/image/feed-test/13.jpg';
-import FeedImg14 from '../../asset/image/feed-test/14.jpg';
-import FeedImg15 from '../../asset/image/feed-test/15.jpg';
-import FeedImg16 from '../../asset/image/feed-test/16.png';
-import FeedImg17 from '../../asset/image/feed-test/17.jpg';
-
-const feedList = [
-  { id: 1, url: FeedImg1, title: '멋사 동료' },
-  { id: 2, url: FeedImg2, title: '멋사 동료' },
-  { id: 3, url: FeedImg3, title: '멋사 동료' },
-  { id: 4, url: FeedImg4, title: '멋사 동료' },
-  { id: 5, url: FeedImg5, title: '멋사 동료' },
-  { id: 6, url: FeedImg6, title: '멋사 동료' },
-  { id: 7, url: FeedImg7, title: '멋사 동료' },
-  { id: 8, url: FeedImg8, title: '멋사 동료' },
-  { id: 9, url: FeedImg9, title: '멋사 동료' },
-  { id: 10, url: FeedImg10, title: '멋사 동료' },
-  { id: 11, url: FeedImg11, title: '멋사 동료' },
-  { id: 12, url: FeedImg12, title: '멋사 동료' },
-  { id: 13, url: FeedImg13, title: '멋사 동료' },
-  { id: 14, url: FeedImg14, title: '멋사 동료' },
-  { id: 15, url: FeedImg15, title: '멋사 동료' },
-  { id: 16, url: FeedImg16, title: '멋사 동료' },
-  { id: 17, url: FeedImg17, title: '멋사 동료' },
-];
-
-export default feedList;
